BUG: Use ProcessObject::GetInput to get generic image
The GrowCut method was using ImageToImageFilter::GetInput method to
access all inputs. This methods returns a pointer to the type of the
first input of the filter. In ITKv3 this pointer was invalid to this
type of object. ITKv4 uses a dynamic_cast to to verify type and return
null if conversion fail. To fix this methods the ProcessObject's
methods is now uses which returns a generic DataObject.