Enable SimpleITK Threads in python, and remove hash version
The SimpleITK external project has been update to use a better
convention of using variable for the repo and git hash. This has also
been updated to refer to my SimpleITK repo with a commit to unlock the
GIL in python. Additionally, for the official SimpleITK repo the git
has is now removed from the version.