WIP: test to check that a volumes origin does not change by viewing
Added python unittest which uses SimpleITK to create a simple image
with a very large origin. It then pushes it to Slicer and views it
with the Volumes module. Currently this changes the values in the
volume's origin.

Unrelated to the illustrated bug, some warnings related to
vtkMRMLVolumeArchetypeStorageNode are displayed, and some times this
test is segfaulting in the other unit test item.

Additionally, this test item was just tacked onto a unit test which is
overly named for a specific action. By reusing this unit test, we are
reusing the Slicer instance, perhaps the python file should be renamed
more generically?